2609_HJ solve

+++ 알고리즘 기초 1/300 - 수학 1/2020.02.17/최소공배수/1934_HJ.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #define MAX_SIZE 1001
 
-long long gcd(long a, long b){
+long long gcd(long long a, long long b){
 	while(b!=0){
 		int r = a%b;
 		a= b;
